EoS missing

+++ rhic/rhic-harness/src/include/edu/osu/rhic/harness/ic/InitialConditionParameters.h
@@ -17,7 +17,7 @@ struct InitialConditionParameters
 	int numberOfNucleonsPerNuclei;
     
     int particleSourceType;//Lipei
-    int initialBaryonDensity;//Lipei
+    double initialBaryonDensity;//Lipei
 
 	double initialEnergyDensity;
 	double scatteringCrossSectionNN;
